// src/app/model/enum/error-messages.enum.ts
export enum ErrorMessagesEnum {
  blocked = "This account has temporarily  blocked!",
  badCredential = "Bad credentials",
  notExist = "This account does not exist!",
  connectRefused = "Server connection error.",
  serverConflict = "Server conflict. Please retry again later!",
  generalError = "Server connection refused."
}